promises/userhandler, test
